import external toolchain support for rpi
This is based on meta-sourcery bfb12523048961042963102f1b1c352438bf453a.
It only installs a "naked" compiler though, libc / libstdc++ are (assumed to be)
provided in the cross sysroot by the deb packages for the target, not from the
cross toolchain. At least in theory they should be compatible when things are sane.

And yes tcmode-external-raspberrypi.inc assumes this is libc6...