attempt to fix the debian packages
For proper string conversion some gconf packages need to be available.
While they are packaged seperatly in OE they don't seem to be in Debian.
So RDEPEND on libc6. Since it is a RDEPEND an not a DEPEND there is no
script figuring out what the exact RDEPEND should be.

https://github.com/jhofstee/meta-bin-deb/issues/2